requirements.txt
accelerate==0.24.0
bertopic==0.16.0
boxprint==0.0.15
rank_bm25==0.2.2
chromadb==0.4.3
cython==0.29.36
datasets==2.14.4
DateTime==5.2
dill==0.3.7
docarray==0.36.0
dvc==2.43.1
dvc-gdrive==2.19.1
feature-engine==1.5.2
hdbscan==0.8.33
jq==1.4.1
jupyter
langchain==0.0.349
lark==1.1.7
llvmlite==0.41.1
lxml==4.9.2
mmh3==3.1.0
multiprocess==0.70.15
nb-black==1.0.7
numba==0.58.1
openai==0.27.8
pip-chill==1.0.1
pinecone-client==2.2.4
pinecone-text==0.7.1
plotly==5.18.0
pre-commit==2.21.0
pyarrow==13.0.0
pydub==0.25.1
pynndescent==0.5.11
pypdf==3.13.0
python-dotenv==1.0.0
qdrant-client==1.7.0
ruff==0.1.4
safetensors==0.3.3
seaborn==0.12.2
sentence-transformers==2.2.2
scikit-learn==1.2.1
spacy==3.5.0
tensorflow==2.13
tiktoken==0.4.0
torch==2.1.1
torchvision==0.16.1
transformers==4.32.0
typeguard==4.1.5
umap-learn==0.5.5
wget==3.2
wikipedia==1.4.0
wordcloud==1.8.2.2
xxhash==3.3.0
yt_dlp==2023.7.6
